pyproject.toml
[project]
name = "handy-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "google-adk>=1.14.1",
]
